# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  entities.CreateUniversityRequest:
    properties:
      name:
        type: string
      postfix:
        type: string
    type: object
  entities.CreateUniversityResponse:
    properties:
      id:
        type: integer
    type: object
...
  entities.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  entities.University:
    properties:
      id:
        type: integer
      name:
        type: string
      postfix:
        type: string
    type: object
...
  entities.UpdateDeleteUniversityResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  title: VUZ+ API
  version: "1.0"
paths:
...
  /auth/university:
    put:
      consumes:
      - application/json
      parameters:
      - description: university data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entities.University'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.UpdateDeleteUniversityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Update university
      tags:
      - university
  /auth/university/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: university id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.UpdateDeleteUniversityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Delete university
      tags:
      - university
  /auth/university/all:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.University'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get all universities
      tags:
      - university
  /auth/university/id/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: university id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.University'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get university by name
      tags:
      - university
  /auth/university/name/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: university name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.University'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get university by name
      tags:
      - university
...
  /auth/work/response:
    get:
      consumes:
      - application/json
      parameters:
      - description: university data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entities.CreateUniversityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CreateUniversityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Response
      tags:
      - work
    post:
      consumes:
      - application/json
      parameters:
      - description: university data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entities.CreateUniversityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CreateUniversityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Response
      tags:
      - work
...
  /university:
    post:
      consumes:
      - application/json
      parameters:
      - description: university data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entities.CreateUniversityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CreateUniversityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Create university
      tags:
      - university
  /work:
    post:
      consumes:
      - application/json
      parameters:
      - description: university data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entities.CreateUniversityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CreateUniversityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Login in hr
      tags:
      - work
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"